Перенес вспомогательные функции filter_by_date, get_date в db_googlesheets.py
Так как возникает ошибка циклического импорта
